\name{fakepackage}
\alias{three}
\alias{is_three}

\title{A fake package.}

\description{This is fake. I'm using it for testing.}
\usage{
three()
is_three(x)
}
\arguments{
  \item{x}{A value.}
}
\details{
  
  \code{three} returns \code{3}. Useful when you need a small, odd
  prime.

  \code{is_three} checks if its argument is three. Not for the faint of
  heart.
}
