% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jed_add.r
\name{jed_add}
\alias{jed_add}
\title{Add two numbers}
\usage{
jed_add(x = 1, y = 1)
}
\arguments{
\item{x}{First number (numeric)}

\item{y}{Second number (numeric)}
}
\value{
The sum of x and y (numeric)
}
\description{
Here is a very short description
}
\details{
Here are some very specific details, like stuff about the inputs.
}
\examples{
jed_add(6,7)

}
