% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jed_sub.r
\name{jed_sub}
\alias{jed_sub}
\title{Subtract two numbers}
\usage{
jed_sub(x = 1, y = 1)
}
\arguments{
\item{x}{First number (numeric)}

\item{y}{Second number (numeric)}
}
\value{
The sum of x and y (numeric)
}
\description{
Here is a very short description
}
\details{
Here are some very specific details, like stuff about the inputs. Even more specific than last time.
}
\examples{
jed_sub(100,44)

}
