% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_bye.R
\name{say_bye}
\alias{say_bye}
\title{Say bye}
\usage{
say_bye(nameToUse = NULL)
}
\arguments{
\item{nameToUse}{Character, the name to wbyech the function says bye.}
}
\description{
Say bye
}
\examples{
say_bye("Luca")
}
