% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hi.R
\name{say_hi}
\alias{say_hi}
\title{Say hi}
\usage{
say_hi(name = NULL)
}
\arguments{
\item{name}{Character, the name to which the function says hi.}
}
\value{
Nothing, outputs a message.
}
\description{
Say hi
}
\examples{
say_hi("Luca")
}
