% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hola.R
\name{say_hola}
\alias{say_hola}
\title{Say hola}
\usage{
say_hola(name = NULL)
}
\arguments{
\item{name}{Character, the name to wholach the function says hola.}
}
\description{
Say hola
}
\examples{
say_hola("Luca")
}
